Bug Fixes and StyleSheet
Added Stylesheet for ChannelChat which defaults it to black background and white fonts.
Set height for badges, which is 1.5x the height of fonts.
If the user's color is the exact same as the background, change it to something else (only have black and grey checking now).
Fixed a bug where system modding tries to access a user that is yet to be added.
Solve #5 by setting a default height to the emotes. It will be changed later to use cache or other methods for emotes with different height